+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix main menu being too small (the other way this time).
 - Add setting to hide focused character.
 - Add skill ruleset selector (Mixed, PvE, PvP).
+- Replace checkboxes with scope slider.
 
 ## 0.5.7
 - Fix main menu being too small.